# docs/api/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crawley analytics API",
    "version": "0.1.0",
    "description": "Presentation JSON for crawley-ui. See docs/api/presentation-v1.md. Worker store paths are not part of this contract."
  },
  "paths": {
...
    "/v1/asx/companies/{ticker}": {
      "get": {
        "tags": [
          "analytics-api"
        ],
        "summary": "Asx Company Detail",
        "operationId": "asx_company_detail_v1_asx_companies__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/asx/scan/start": {
      "post": {
        "tags": [
          "analytics-api"
        ],
        "summary": "Asx Scan Start",
        "operationId": "asx_scan_start_v1_asx_scan_start_post",
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Re-run even if active set already scanned",
              "default": false,
              "title": "Force"
            },
            "description": "Re-run even if active set already scanned"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "analytics-api"
        ],
        "summary": "Jobs Get",
        "operationId": "jobs_get_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/asx/portfolio/trade": {
      "post": {
        "tags": [
          "asx-extra-api"
        ],
        "summary": "Asx Portfolio Trade",
        "operationId": "asx_portfolio_trade_v1_asx_portfolio_trade_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperTradeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Asx Portfolio Trade V1 Asx Portfolio Trade Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/asx/companies/{ticker}/notebook": {
      "get": {
        "tags": [
          "asx-extra-api"
        ],
        "summary": "Asx Notebook Get",
        "operationId": "asx_notebook_get_v1_asx_companies__ticker__notebook_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Asx Notebook Get V1 Asx Companies  Ticker  Notebook Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "asx-extra-api"
        ],
        "summary": "Asx Notebook Put",
        "operationId": "asx_notebook_put_v1_asx_companies__ticker__notebook_put",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotebookBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Asx Notebook Put V1 Asx Companies  Ticker  Notebook Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyDetailResponse": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "scan_status": {
            "type": "string",
            "title": "Scan Status",
            "default": "pending"
          },
          "snapshot": {
            "$ref": "#/components/schemas/SnapshotDTO"
          },
          "move": {
            "type": "string",
            "title": "Move",
            "default": "\u2014"
          },
          "headlines": {
            "items": {
              "$ref": "#/components/schemas/HeadlineDTO"
            },
            "type": "array",
            "title": "Headlines"
          },
          "sources_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources Used"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileDTO"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "Not licensed research \u2014 operator notes and desk synthesis only."
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "name",
          "sector",
          "snapshot",
          "profile"
        ],
        "title": "CompanyDetailResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HeadlineDTO": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "url": {
            "type": "string",
            "title": "Url",
            "default": ""
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": ""
          }
        },
        "type": "object",
        "title": "HeadlineDTO"
      },
...
      "JobProgressDTO": {
        "properties": {
          "processed": {
            "type": "integer",
            "title": "Processed",
            "default": 0
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining",
            "default": 0
          },
          "current_ticker": {
            "type": "string",
            "title": "Current Ticker",
            "default": ""
          }
        },
        "type": "object",
        "title": "JobProgressDTO"
      },
      "JobStatusResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "idle",
              "busy",
              "paused",
              "done",
              "error"
            ],
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "error": {
            "type": "string",
            "title": "Error",
            "default": ""
          },
          "progress": {
            "$ref": "#/components/schemas/JobProgressDTO"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "default": ""
          },
          "pause_requested": {
            "type": "boolean",
            "title": "Pause Requested",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "progress"
        ],
        "title": "JobStatusResponse"
      },
...
      "NotebookBody": {
        "properties": {
          "thesis": {
            "type": "string",
            "title": "Thesis",
            "default": ""
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "title": "NotebookBody"
      },
      "PaperTradeBody": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "side": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ],
            "title": "Side",
            "default": "buy"
          },
          "qty": {
            "type": "number",
            "title": "Qty",
            "default": 1.0
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "ticker"
        ],
        "title": "PaperTradeBody"
      },
      "ProfileDTO": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "empty"
          },
          "markdown": {
            "type": "string",
            "title": "Markdown",
            "default": ""
          },
          "error": {
            "type": "string",
            "title": "Error",
            "default": ""
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "default": ""
          }
        },
        "type": "object",
        "title": "ProfileDTO"
      },
...
      "ScanActionResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "job": {
            "$ref": "#/components/schemas/JobStatusResponse"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "message",
          "job"
        ],
        "title": "ScanActionResponse"
      },
      "SnapshotDTO": {
        "properties": {
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "change_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Pct"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "AUD"
          },
          "sentiment": {
            "type": "string",
            "title": "Sentiment",
            "default": "unknown"
          },
          "headline": {
            "type": "string",
            "title": "Headline",
            "default": ""
          },
          "as_of": {
            "type": "string",
            "title": "As Of",
            "default": ""
          },
          "gaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Gaps"
          }
        },
        "type": "object",
        "title": "SnapshotDTO"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}